backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.20  # Required for file upload support in FastAPI
google-generativeai
python-dotenv
google-genai==1.53.0
